backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn==0.29.0

# Data & ML
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
imbalanced-learn==0.12.0

# Visualisation (Plotly only — matplotlib/seaborn/kaleido not used)
plotly==5.20.0

# File upload
python-multipart==0.0.9

# AI providers
google-genai
groq

# Environment / config
python-dotenv
